[3.0] Fix OpenGL type-mappings (#1907)
* Initial untested work on the cache mechanism

* Fix dependency injection bugs

* Minor cleanups

* Make unit test representative

* Fix the caching problem

* Regenerate Windows bindings

* Adjust git attributes

* Flush the file stream maybe?

* Delete just in case gitattributes didn't apply

* Readd .stout files

* Fix some pathing schenanigans

* Regenerate bindings

* Apply generalisation to the whole cache key

* That's relativity folks

* Just hash the original input rather than recreating it

* Crank up logging

* Another attempt at the cache key

* Another attempt at the cache key

* Regenerated w/ cache finally!

* Small cleanups now that caching is working

* Fix OpenGL type mappings

* Delete non-existent files in the sln

* Add generator.json to sln

* Remove test.json

---------

+++ sources/Playground/Program.cs
@@ -41,7 +41,7 @@ unsafe
     GL.BindBuffer(GL.Constants.ArrayBuffer, vbo);
     GL.BufferData(
         GL.Constants.ArrayBuffer,
-        vertices.Length * sizeof(float),
+        (nuint)(vertices.Length * sizeof(float)),
         vertices[0].AsRef(), // TODO simplify this
         GL.Constants.StaticDraw
     );
